Add automatically configure hardware acceleration Add pipewire Asus motherboard check

+++ scripts/pipewire.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
 arch-chroot /mnt /bin/bash -c "pacman -S --needed --noconfirm alsa-utils alsa-firmware alsa-card-profiles alsa-plugins pipewire pipewire-pulse pipewire-alsa pipewire-jack"
-arch-chroot /mnt /bin/bash -c "mkdir -p /etc/pipewire"
-cp -rf ../tweaks/pipewire/* /mnt/etc/pipewire
+if aplay -L | grep -q ALC887 ; then
+  arch-chroot /mnt /bin/bash -c "mkdir -p /etc/pipewire"
+  cp -rf ../tweaks/pipewire/* /mnt/etc/pipewire
+fi
+
